requirements.txt
python-pptx>=0.6.21
openai>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-docx>=1.1.0
pypdf>=4.0.0
pydantic>=2.12.0
pytest>=8.0.0
